makes live tserver set use keep a precomputed snapshot
LiveTseverSet is modified to keep a snapshot that is only recomputed on
changes.  It used to recompute for each request and it also used to keep
multiple indexes up to date.  The indexes are no longer kept up to date
but instead are recomputed on demand which simplified the code in
addition to making it more efficient.