Refactors tablet mgmt iterator to share decision code w/ TGW
The TabletGroupWatcher uses the TabletManagementIterator to filter
tablets that need attention.  The TabletMgmtIter had its own custom code
to make decision about tablets.  This commit modifies the
TabletMgmtIterator to use the same code as the TGW when making decisions
about tablets.  This makes it easier to reason about and change the behavior
of the TGW.

In making the TGW and TabletMgmtIter share code a change was also made
to how they get information to make decisions.  A new class called
TabletManagementParameters was introduced that contains an immutable
snapshot of the information that both classes need to make decisions.
With this change the iterator and TGW are using the same information for
a pass over the metadata table, in the past the information was obtained
at different times and could have been different due to race conditions.
These race conditions were probably not harmful, but removing them makes
the code easier to reason about.

Also move some code outside of TGW away from using the
TableMgmtIterator. Need to move all code outside of TGW away from using
this iterator in order to make the code easier to maintain.  Left TODOs
in the code about this and will open follow on issues.