Fix inconsistent view of TabletServers in Manager (#3901)
Modified LiveTserverSet so that the set of tablet servers and the tablet
servers resource groups are acquired atomically.  The code was
acquiring this information at two different times with two different lock
acquisitions, which could have led to race condtions resulting
differences in set and the map. Fixed some cases in the balancers
that were not handling an empty or non-existent group correctly.


Co-authored-by: Keith Turner <kturner@apache.org>