Fixes WaitForBalanceIT
It appears that Manager.waitForBalance may return and then the
isBalanced method returns false. I think this is due to the fact
that the current location for the Tablet is set in the TabletServer
AssignmentHandler, which leaves a gap where
InstanceOperations.waitForBalance returns but the locations are not
yet set.