Fixes handling of incorrect compaction configuration
An incorrectly configured compaction service would fail to create.  This
would cause tables configured to use that service to spam the logs
saying the serivce did not exists.  Two changes were made to address
this.

First, when a compaction service fails to create its planner plugin it
will log an error and fall back to using a new NullPlanner that does
nothing.  Once the configuration is fixed and the planner pluging can be
created, it will replace the NullPlanner. Falling back to the
NullPlanner allows the service to exists and do nothing, this way tables
configured to use it do not complain.

Second, when a table is configured to use a compaction service that does
not exist repeated logging of this is suppressed per table.  Also the
existing code attempted to fall back to the default compaction service
but this was not working.  So the fallback code was removed and the log
message was adjusted.  Falling back could cause problems if its not what
the user desires.

Two ITs were added to test the above situations.