WIP adds support for splitting tablets with walogs
Updates the FATE operations that split tablets to handle walogs.  These
changes should wait for a tablet with walogs to recover before
proceeding.  The changes to actually make recovery happen were already
done in #3904 with changes to TabletGoalState.compute().  This change is
a WIP because it depends on #3847 and needs those changes to be
complete.

fixes #3844