Replace logic for determining whether a Tablet is in use (#3920)
The previous logic used the ingest and scan rates and would
say a Tablet was not in use when the rates were zero. However,
due to a smoothing function on the Rate class, this may not
happen or would take a really long time. The logic was modified
to use the internal Tablet activeScans and writesInProgress
variables to determine if activity was currently happening
instead of using the Rate-based metrics.

Fixes #3907


Co-authored-by: Keith Turner <kturner@apache.org>