updates delete table to use conditional muations (#3929)
Updates the delete table fate operation to use conditional mutations.
Also introduces a new mechanism in Ample for processing conditional
mutations that does not buffer everything in memory.  This avoid
buffering large amounts of tablet metadata in memory when deleting a
large table.  The IT that creates 1 million splits in a table and
deletes it passes with these changes.