Updates merge code to support merging many tablets
Using the conditional writer changes from #3929 this commit updates the
merge code to support merging more tablets than will fit in memory. Also
adds merge and clone operations to the SplitMillionIT and merges a table
with one million tablets in the test.  Without the changes in this
commit that test would cause the manager to die with an out of memory
error.