Adds check for merging too many files
A check was added to ensure that tablets being merged will not result in
creating a tablet with a massive number of files.  Added test for
merging tablets with too many files.  Removed the code in merge that
wrote files in multiple mutations.  Went back to buffering all files and
writing them in a single mutation.