Updates merge code to support merging many tablets (#3934)
Using the conditional writer changes from #3929 this commit updates the
merge code to support merging more tablets than will fit in memory. Also
adds merge and clone operations to the SplitMillionIT and merges a table
with one million tablets in the test.  Without the changes in this
commit that test would cause the manager to die with an out of memory
error.

Added a check to ensure the number of files in the merge range does not 
exceed a configurable maximum. With these changes its possible to merge 
lots (like millions) of empty tablets.  However if merging lots of tablets with 
files, then the new check will trigger to avoid creating a tablet with too many
files. Added test for merging tablets with too many files.