adds support for compacting lots of tablets
Removes buffering of all conditional results in memory in the compaction
code.  This allows compacting more tablets than would fit in memory.

Changed the condition result BiConsumer to a Consumer because it was
passing the extents twice, so a BiConsumer was not needed and made the
code more verbose.

Updated the compaction code to collect stats and log a trace.  It was
logging a lot of per tablet information at debug.  When compacting one
million tablets, this resulted in a lot of information in the manager
logs.  Moved the per tablet information to trace logging.  Added
collection of stats for the different per tablet information and logged
the stats once for all tablets scanned.

Added compaction to the SplitMillionIT. Without the other changes in
this PR adding compaction to the SplitMillion would cause the Manager
to die with an out of memory error because the conditional write was
buffering all 1 million tablets.