Cleanup failed compaction tmp files
Modified TabletNameGenerator.getNextDataFilenameForMajc to add
the external compaction id to the filename. When a compaction
fails, or the DeadCompactionDetector notices a Compactor is down,
CompactionCoordinator.compactionFailed is called to remove the
external compaction id from the tablet metadata. Modified this
method to also remove corresponding tmp files from the tablet
directory when the external compaction id is successfully removed
from the tablet metadata. Created the FindCompactionTmpFiles
utility to find, and optionally delete, compaction tmp files
at a system level. This utility requires that no Compactors are
running to avoid any collisions.

Fixes #3577