Replaces java constant name with value in metrics javadocs
The metrics javadocs, which are referenced from the user manual, have a
table that currently contains java constant names.  From the perspective
of a user they would probably prefer the value of the java constant,
which is currently cumbersome to obtain. This commit replaces the java
constant name with the values in the javadoc.