fixes bug with compaction preemption
User compactions can preempt selector compaction under certain
conditions.  There was a bug in the code where a user compaction could
preempt a selection compaction that was in the middle of selecting its
files.  The way the code is structured doing this causes an exception.
This change makes the preemption wait until file selection is complete.