Cleanup failed compaction tmp files (#3955)
Modified TabletNameGenerator.getNextDataFilenameForMajc to add
the external compaction id to the filename. When a compaction
fails, or the DeadCompactionDetector notices a Compactor is down,
CompactionCoordinator.compactionFailed is called to remove the
external compaction id from the tablet metadata. Modified this
method to also remove corresponding tmp files from the tablet
directory when the external compaction id is successfully removed
from the tablet metadata. When the external compaction id is not
successfully removed, like when the tablet is merged or deleted,
then the cleanup of table tmp files will occur in the background
DeadCompactionDetector thread. Created the FindCompactionTmpFiles
utility to find, and optionally delete, compaction tmp files
at a system level.

Fixes #3577