refactors upgrade to use same code for all tablet metadata
The upgrade code called different code depedning on where tablet
metadata was stored.  This change updates the code to use the same code
to process all tabletmetadata wether its stored in ZK or an accumulo
table.

A secondary purpose of this changes is pull all upgrade code into the
Upgrader11to12 class so that when that class is eventually deleted, no
lingering code is left in the code base related to that specific
upgrade.

RootTabletMetadataTest was deleted and some of its test were migrated
to Upgrader11to12Test