fixes bug with bulk import RPC failure and retry
The bulk import code had a bug where if an exception was thrown in a
certain place in the tablet code that it would mark the bulk import as
complete.  After the exception, other code would retry and see the bulk
import was complete even though it was not.  This commit changes the
behavior to only note success when there is no exception. In addition
test were added that recreated this bug.  These test set a constraint on
the metadata table that cause bulk import writes to fail.