Deprecation of Internal Compaction code
This change deprecates the internal compaction feature in accumulo and
also introduces classes that will help to migrate off the extensive use
of "external" once the internal compaction feature is fully removed in a
later version.

Along with this deprecation, the concept of a "compaction group" has
been added. For now, this is not a functional change, but just a different
wording regarding the external compaction queues.