introduces interface for compaction coordination
As an initial step for #4005, introduces an interface for the compaction
coordination service in the manager.  Everything in the manager uses
this interface to interact with the compaction coordinator.  In order to
make this change metrics and compaction queues were moved from manager
code to the compaction coordinator code.  Both of these changes were
good for encapsulation.   Hopefully further changes for #4005 lead to
more encapsulation and less code in the Manager class.

fix bug