Update delete rows to be idempotent by always keeping at least 1 tablet (#4008)
This fixes an edge case bug where delete rows could fail to recover if
all tablets are removed during delete and the operation fails before a
new default tablet can be created which will prevent the fate operatino
from resuming. Instead, now if the last tablet is included in the delete
range (null end row) we just keep the tablet and only remove the files
and update that tablets prev end row instead of creating a new tablet
which will guarantee we always have at least 1 tablet remaining and we
can recover the operation if it fails at any point.