refactor fates storage layer (#4017)
Refactored the storage layer of FATE to return an object when reserving
a fate transaction.  This object allows mutating the storage related to
that FATE transaction.  This replaces methods where the fate transaction
id had to to always be passed.

These changes are a subset of the changes in #3964, but only focusing on
the storage layer refactoring and nothing else.

Co-authored-by: Christopher L. Shannon (cshannon) <christopher.l.shannon@gmail.com>