modifies FATE to use a single thread to find work (#4042)
This change modifies FATE to use singe thread to find work.  It also
cleans up some of the signaling between threads in FATE and fixes a
synchronization bug in FATE that was introduced in #4017.  The bug
introduced in #4017 is that somethings are syncronizing on the wrong
object because a new inner class was introduced.

These changes were pulled from #3964 and cleaned up and improved.