Remove redundant LogEntry value (#4071)
Remove the `LogEntry.getValue()` method and use the column qualifier
for the file path, without the redundant storage in the Value

* Remove `getValue()` method and use empty `new Value()`
* Create addToMutation method to remove boilerplate for storing
  the LogEntry in the metadata
* Strictly validate the metadata log entries' column qualifier

This re-applies the reverted change for #3997,
0cdf6dd2de0ca382c3881169bfe90c3bbd4367b3, but without the renaming of
filePath to logReference, since with further analysis of the schema, it
was discovered that these are, in fact, file paths, and not just generic
references to log IDs.

This re-application also removes the metadata constraint that expected
the value to be non-empty, which was overlooked on the first attempt.

---------

Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>