Fix MemoryStarved ITs
In #3951 the ScanServer and Compactor advertise the Thrift Client
service. Since then, the freeServerMemory method would connect
to any TabletServer, Compactor, or ScanServer during the test.
However, we need it to only connect to the TabletServer. Modified
freeServerMemory to only connect to the TabletServer and introduced
a wait in MemoryStarvedMajCIT as it was trying to connect before
the Compactor was fully up.