Fixes FATE deferral
This fixes a bug where when a FATE operation was not ready and was
deferred that it may never run again.  Also improved some logging that
was encountered while debugging this.  This bug was causing test in
ConcurrentDeleteTableIT to hang forever because concurrent fate
operations were deferred and then never ran again.