Avoid DfsLogger unencrypted WAL special case (#4102)
Although the NoCryptoService's decrypter returns the same stream that it
was given in the current implementation, that behavior is not guaranteed
in the future. So, instead of relying on that insider knowledge about
the NoCryptoService implementation, ask the decrypter for the stream
like normal, and then check if it's already a DataInputStream before
re-wrapping it, since that's the condition we actually should check in
order to avoid rewrapping. This allows us to avoid duplicate wrapping of
other implementations as well, instead of treating NoCryptoService as a
special case.