Updates DefaultCompactionPlanner to honor table.file.max prop (#4127)
This change updates the DefaultCompactionPlanner to honor the
table.file.max property.  When a tablet has more files than is
configured in table.file.max and its not compacting, it will try to
find a compaction ratio that will cause a compaction.

Using a lower compaction ratio will find files of a similar size to
compact. This avoids the problem with deprecated DefautlCompactionStrategy
where it may select a 1K and 1G file for compaction because it selected
the N smallest files that would bring the tablet under the
table.file.max limit.