Add support for multiple fate instances
This commit updates the Manager to create two fate instances, one for
store FATE operations against tables in the accumulo namespace (meta) in
the existing zookeeper store and another fate instance to store FATE
operations against user tables (user). The thrift API has been modified
so that when acquiring a new transaction, the type is returned of the
instance used because future operations with the same transaction will
need to know which instance was used.