Moves split recovery code to upgrade
Splits write three mutations to the metadata table. In the case of
process death all writes may not happen.  Splits now run in FATE so
this is not a concern, but did not previously run in FATE.  There was
custom code to handle this failure case that wrote specialized
columns to the metadata table. Handling of these specialized columns
was moved into the upgrade code.

Existing test were adpated to test the moved code.  One test was removed
because it was no longer relavant.  The test simulated split recovery on
tablet load which is something that no longer happens, so the test was
removed.