removes consitency check in GC wal scan
The GC Wal scan was doing consitency checking and filtering which caused
a runtime excecption that prevented wal GC.  Consistency checking is not
needed becasue a tablet with a wal will not split or merge. Prior to the
change in 5c454e12a144e3543175d2cbc0b54bc0900749b1 consistency checking
was not done.  It was added w/o a specific reason so removing instead of
removing filtering. Fixing GC for wals fixes VolumeIT which was failing
because no GC was happening.