fixes split recovery for metadata tablet (#4137)
When a metadata tablet splits it writes to the root table.  The code
that handles split recovery was always writing to the metadata table.
Therefore if a tablet server was in the middle of splitting a metadata
tablet and died, then it would not recover correctly.  This change
modifies the split recovery code to use the appropriate table.

Noticed this while working on #4136