Add support for multiple Fate instances to the Manager (#4133)
This commit updates the Manager to create two fate instances, one for
store FATE operations against tables in the accumulo namespace (meta) in
the existing zookeeper store and another fate instance to store FATE
operations against user tables (user). The thrift API has been modified
so that when acquiring a new transaction, the type is returned of the
instance used because future operations with the same transaction will
need to know which instance was used.

The Fate Admin command needs to be updated now that multiple instance
types are used for Fate operations. This commit fixes the ITs that call
the admin command operations but there is still more work to be done to
complete all the changes.

Co-authored-by: Keith Turner <kturner@apache.org>