Replaces BulkFailureIT with new test
BulkFailureIT excercised low level bulk import implementation details
that no longer exists in the elasticity branch.  The test was removed
and a new test was added that operates at a much higher level.  The new
test runs concurrent bulk import and compactions in such a way that
duplicate loads of a file could be detected.