Exposes status when listing fate operations (#4169)
The Accumulo store can very efficiently gather status the status at the
same time as the id for a FATE transaction.  This is already being
gathered internally in implementation of the fate storage layer.
This commit exposes this information further as it will be useful in
refactoring the ageoff store.