Avoids buffering large split points in memory
When a tablet needs to split a set split points is found.  These split
points must be under a desired threshold.  The code that found these
would buffer everything in memory and then remove any that were over the
threshold.  This commit avoid buffering anything that is over the
threshold in the first place.

Noticed this behaviour while debugging #4172