Avoids buffering large split points in memory (#4173)
When a tablet needs to split a set of split points is found. These split points must be under a desired threshold. The code that found these would buffer everything in memory and then remove any that were over the threshold. This commit avoids buffering anything that is over the threshold in the first place.

Noticed this behavior while debugging #4172