WIP moves finding split points into FATE
 1. Moves finding split points from tablet from a custom thread pool in
    the manager into FATE.  This is done by the new Repo named FindSplits.
 2. Stops tracking what splits are running in manager memory and moves
    this into FATE.  This is not implemented yet, plan to do that in
    another PR
 3. Stops tracking what tablets are unsplittable in manager memory. This
    should be tracked in metadata table per #4177.  This change can be
    in its own commit.  In this commit the in memory tracking is simply
    removed.

These changes will reduce manager memory usage. These changes also will
make splitting tablets work much better if FATE is distributed for two
reasons.  First it allows the computation to spread over multiple
manager processes.  Second it allows tracking of what is splitting to
move from memory to persisted storage and not rely on a map in a single
process.

This is a WIP commit because it requires FATE changes.  The needed FATE
changes were stubbed out in this PR and implemented with a in memory set
that would need to be persisted.