Fix fateRefs init check when setting the manager state
As part of #4133 the null check for fateRef inside of setManagerState()
method in the Manager was inadvertently changed to check if the map size
was non empty but the actual check should be that it's null which means
not initialized. The map will never be non empty if it is set.

This restores the previous check that used to check for a null ref