Fixed TabletManagmentParameters serialization of replacements
TabletManagementParameters was still using a List<Pair<Path,Path>>
for volume replacements serialization. This was changed to a
Map<Path,Path> in other places of the code because there was an
issue with Gson serialization. I added a test for serialization
and deserialization and was able to successfully use a Map when
setting enableComplexMapKeySerialization() on the Gson object.

Fixes #4107