Initial commit for issue 4044
A WIP for #4044. The end goal is to have the stronger type FateId replace the current representation of a transaction id (which is just a long). This was brought about from the addition of the AccumuloStore class - there are now two fate instance types associated with a transaction - META (for ZooStore) or USER (for AccumuloStore). FateId is a new class which includes the FateInstanceType and the transaction id.
Current changes:
- FateTxId replaced with new class FateId
- Started with changes in ReadOnlyFateStore and resolved the issues in other classes extending from changing this class.
Still left TODO:
- There are still related problems from the classes I have changed: I have not yet finished going down the entire chain of issues
- Need to change Fate and the associated issues with changing this class
- Need to change Repo and the associated issues with changing this class
- More changes TBD