Fixes RowFilter bug with deep copy
If a RowFilter implemented init and used the options, then these options
would be lost if the RowFilter was deep copied.  When RowFilters are
chained it will the 2nd RowFilter in the chain will deep copy the first.
This commit fixes the bug by saving options and calling init with them
in deep copy.