Add support for creating a FATE transaction by key
This updates FATE to support creating a transaction by a key which will
be hashed to a transaction id. This can be used to guarantee only one
transaction exists at the same time if the key is the same. Currently
the key is just a byte array so anything can be used for the key.

This closes #4183