Refactor AccumuloStoreReadWriteIT to also run against ZooStore
This renames AccumuloStoreReadWriteIT to FateStoreIT and adds an
implementation for both the Accumulo and Zookeeper Fate stores. The
tests in this class apply to both so this change allows all the tests to
run against both implementations

The helper executor methods that were part of FateIT were moved into
FateTestRunner to be re-used between the FateIT and FateStoreIT tests.