Fix intermittent failure in AuthenticationTokenSecretManagerTest (#4203)
There fixes a test race condition in the testVerifyPassword test that
could occasionally cause the test to fail if it ran too quickly. Token
generation uses System.currentTimeMillis() so the test previously would
fail when trying to generate a new token if the timestamp didn't update
as the same token would be returned. This commit adds a short sleep of 50
ms to guarantee that the test won't fail.