Host tablets needing recovery (#4229)
Modified the TabletManagementIterator to return NEEDS_RECOVERY
when the tablet has wals and is not being deleted. Modified
TabletGoalState to return the goal of HOSTED when the tablet
has wals and tablet availability is UNHOSTED or ONDEMAND. Re-ordered
the TabletGroupWatcher such that volume replacement is always
performed. In the case that bad state is not returned, but recovery is,
then recovery will be performed.

Closes #3663