Adjusts compactor error logging because of race (#4151)
Noticed error logging from the compactor process and found it was caused
by a race condition.  The logging indicated a compaction was not
running when it was expected to, but the compaction started shortly
after.  The background thread had not yet populated a set that the
foreground thread was examining.  Adjusted the error logging to debug
since it can happen spuriously.