Fixes GC WAL bug (#4196)
This fixes a GC bug that only exists in the elastictiy branch. When
garbage collecting write ahead logs the metadata table should be
inspected to look for tablets with WALs and tablets assigned to dead
tablet servers.  The GC code was only looking for tablets with WALs.
This commits makes it also look for tablets assigned to dead
tservers.

This bug was noticed because CleanWalIT was sometimes failing.
Sometimes the GC was collecting a WAL after the metadata table recovered
but before user tablets could be assigned the WAL.  This caused the user
tablets to silently loose data.  The test caught this data loss.

In the course of fixing this bug a change was made to how RowFilter was
used and a pre-existing bug with RowFilter was found and fixed.  This
bug fix will be back ported separately.