Add support for creating a FATE transaction by key (#4204)
This updates FATE to support creating a transaction by a key which will
be hashed to a transaction id. This can be used to guarantee only one
transaction exists at the same time if the key is the same.  Currently,
two types of keys exist for split and compaction operations. A
new method has been added, createAndReserve(fateKey) which
allows creating and reserving the new transaction to prevent race
conditions. This commit also updates the ZK store so FateKey 
storage is saved to the node for the ZooStore so it will be inserted
atomically with the transaction id.

This closes #4183

Co-authored-by: Keith Turner <kturner@apache.org>