fixes volume replacement failures (#4257)
VolumeIT was failing because of two problems with volume replacement.
First, tablets needing volume replacement were being assigned preventing
volume replacement from happening. A change was made to TabletGoalState
to fix this.  Second when tablet was assigned to dead server, the
location of the dead server was never removed preventing volume
replacement.  A change was made to TabletManagementIterator to fix this.