Fixes CompactionIT.testMultiStepCompactionThatDeletesAll() (#4259)
CompactionIT.countFiles() was counting in an overly broad way.  In the
elasticity branch that included the new +fate table and the new availability
columns in ~tab. This was causing a test failure. Narrowed the scope of
countFiles to fix this.  Even though the test was not broken in 2.1, the
test fixes were made there because its a general improvement to the test.