Resolved timeout issue in MemoryStarvedMajCIT (#4267)
It appears that the test was stopping the Compactors, but not
waiting for their ZooKeeper locks to be removed, before moving
on and checking the number of zookeeper locks. I think what was
happening is that the test code moved forward thinking that the
Compactors were stopped, but the Compactors were actually stopping
after this point and during the next step in the test.