Fixed MemoryStarvedScanIT timeout (#4270)
MemoryStarvedScanIT was timing out because the IT was deleting
tables as part of the test and scans on the new FATE table were
being blocked because of the low memory condition. Adjusted the
check for whether the table is a user table or not to include
all tables in the accumulo namespace.