fixes FateStarvationIT
This test had two problems.  First it was using the offline operation to
check if all compactions were finished.  However compactations that were
queued in fate when the offline started would start after the offline
and potentially have fate locks for the check at the end of the test.
Second the test starts a large number of compactions, but there was only
one compactor process.  Started multiple compactor process to process
compactions faster.