avoids a metadata lookup per mutation for batch writer
When writing to unhosted ondemand tablets the batch writer was peforming
a metadata lookup per mutation before it requested hosting.  This was of
course extremely slow.  This commit makes a quick change to perform a
metadata lookup per extent instead of per mutation.  This could be
futher improved because its still not as good as the pre elasticity
code. There is in ELASTICITY_TODO in the commit about further
improvements.

This is a massive improvement.  Was running FateStarvationIT which
writes 100K rows.  Without this change it took 53 secs.  With this
change it took 1.7 secs.

This mostly solves #3708, there may be some futher minor improvements
that could be made.