persist time to tablet in bulk update (#4072)
When bulk import operations set time and a tablet was hosted the time
was not persisted.  The bulk import fate operation now persist time in
tablet metadata.  The tablet code assumed it was the only thing
updating a tablets time field.  The tablet code was modified to
accomodate the bulk import code running in the manager updating the
tablets time column in the metadata table.